Add comments, use attribute accessors for content serialization
Instead of calling `getSortedContents` and manually constructing arrays
from those results, we can just use Eloquent ORM's attribute and
serialization features to automatically append a `contents` field to the
results of `toArray` that contain the sorted contents of a board. This
way, we can also just call `toArray` on a Folder or Word and we'll
automatically get sorted `contents` if the Model class implements an
`$appends` property. Neat!