Fix bug for manual insertion of auto-incrementing ids
When normally creating models, we omit the `id` field and the DB just
generates the id since the DB column is SERIAL or auto-incrementing in
some other way.

When we manually pass in the `id` field, like in `BoardSeeder`, the DB
doesn't update what the next auto-increment should be. Because of this,
when we try to create a word the usual way (omitting the `id` field)
after running the seed script, the DB tries to generate an id of 1.
We've already inserted a row w/ id 1 in the seed script, so the DB blows
up.

This just manually resets the next auto-increment value to whatever is
the last id we manually inserted. Hopefully this doesn't pop up again,
this was a nasty one to debug.